# JDK-8312182: THPs cause huge RSS due to thread start timing issue