--- a/src/hotspot/os/linux/os_linux.cpp
+++ b/src/hotspot/os/linux/os_linux.cpp
@@ -916,22 +916,12 @@ bool os::create_thread(Thread* thread, ThreadType thr_type,
   size_t stack_size = os::Posix::get_initial_stack_size(thr_type, req_stack_size);
   size_t guard_size = os::Linux::default_guard_size(thr_type);
 
-  // If THPs are unconditionally enabled, the following scenario can lead to huge RSS:
-  // - parent thread spawns, in quick succession, multiple child threads
-  // - child threads are slow to start
-  // - thread stacks of future child threads are adjacent and get merged into one large VMA
-  //   by the kernel, and subsequently transformed into huge pages by khugepaged
-  // - child threads come up, place JVM guard pages, thus splinter the large VMA, splinter
-  //   the huge pages into many (still paged-in) small pages.
-  // The result of that sequence are thread stacks that are fully paged-in even though the
-  // threads did not even start yet.
-  // We prevent that by letting the glibc allocate a guard page, which causes a VMA with different
-  // permission bits to separate two ajacent thread stacks and therefore prevent merging stacks
-  // into one VMA.
   if (PreventTHPsForThreadStacks) {
-    guard_size = MAX2(guard_size, os::vm_page_size());
-    // Add an additional page to the stack size to reduce its chances of getting huge page aligned
-    // so that the stack does not get backed by a transparent huge page.
+    // In addition to the glibc guard page that prevents inter-thread-stack hugepage
+    // coalescation (see comment in os::Linux::default_guard_size()), we also make
+    // sure the stack size itself is not huge-page-size aligned; that makes it much
+    // more likely for thread stack boundaries to be unaligned as well and hence
+    // protects thread stacks from being targeted by khugepaged.
     if (HugePages::thp_pagesize() > 0 &&
         is_aligned(stack_size, HugePages::thp_pagesize())) {
       stack_size += os::vm_page_size();
@@ -3095,6 +3085,27 @@ bool os::Linux::libnuma_init() {
 }
 
 size_t os::Linux::default_guard_size(os::ThreadType thr_type) {
+
+  if (PreventTHPsForThreadStacks) {
+    // If THPs are unconditionally enabled, the following scenario can lead to huge RSS
+    // - parent thread spawns, in quick succession, multiple child threads
+    // - child threads are slow to start
+    // - thread stacks of future child threads are adjacent and get merged into one large VMA
+    //   by the kernel, and subsequently transformed into huge pages by khugepaged
+    // - child threads come up, place JVM guard pages, thus splinter the large VMA, splinter
+    //   the huge pages into many (still paged-in) small pages.
+    // The result of that sequence are thread stacks that are fully paged-in even though the
+    // threads did not even start yet.
+    // We prevent that by letting the glibc allocate a guard page, which causes a VMA with different
+    // permission bits to separate two ajacent thread stacks and therefore prevent merging stacks
+    // into one VMA.
+    //
+    // Yes, this means we have two guard sections - the glibc and the JVM one - per thread. But the
+    // cost for that one extra protected page is dwarfed a large win in performance and memory that
+    // avoiding interference by khugepaged buys us.
+    return os::vm_page_size();
+  }
+
   // Creating guard page is very expensive. Java thread has HotSpot
   // guard pages, only enable glibc guard page for non-Java threads.
   // (Remember: compiler thread is a Java thread, too!)