@tstuefe 2023-07-19
@poonamparhar @dholmes-ora Thanks for your respective feedback. New version slightly moves the guard page calculation around and improves comment, functionally the same.