@tstuefe 2023-07-20
New version:
- changed flag name from `PreventTHPsForThreadStacks` -> `DisableTHPStackMitigation` (inverse logic) and made it diagnostic
- comment corrections
- changed the negative tests to be a jtreg / manual test (@dholmes-ora is that the right way to go here? I want a test that does not run automatically, but that can be run from command line)